Rename to `TickerBucketImportExportProvider` for consistency
