+++ src/hooks/useBucketImportExportContext.ts
@@ -1,12 +1,12 @@
 import { useContext } from "react";
 
-import { BucketImportExportContext } from "@providers/BucketImportExportProvider";
+import { BucketImportExportContext } from "@providers/TickerBucketImportExportProvider";
 
 export default function useBucketImportExportContext() {
   const context = useContext(BucketImportExportContext);
   if (!context) {
     throw new Error(
-      "useBucketImportExport must be used within a BucketImportExportProvider",
+      "useBucketImportExport must be used within a TickerBucketImportExportProvider",
     );
   }
   return context;
