+++ src/router.tsx
@@ -7,7 +7,7 @@ import PortfoliosPage from "@pages/Portfolios.page";
 import SearchResultsPage from "@pages/SearchResults.page";
 import SettingsPage from "@pages/Settings.page";
 import WatchlistsPage from "@pages/Watchlists.page";
-import BucketImportExportProvider from "@providers/BucketImportExportProvider";
+import TickerBucketImportExportProvider from "@providers/TickerBucketImportExportProvider";
 import { RouteObject, createBrowserRouter } from "react-router-dom";
 
 import MainLayout from "@components/MainLayout";
@@ -20,9 +20,9 @@ export const routes: RouteObject[] = [
       // Note: `SharedSessionManagerProvider` is placed here instead of in `App`
       // due to being dependent on `React Router`.
       <SharedSessionManagerProvider>
-        <BucketImportExportProvider>
+        <TickerBucketImportExportProvider>
           <MainLayout />
-        </BucketImportExportProvider>
+        </TickerBucketImportExportProvider>
       </SharedSessionManagerProvider>
     ),
     children: [
